+++ scripts/helperFunctions.js
@@ -232,14 +232,19 @@ export let chris = {
 	'functionToString': function _functiongToString(input) {
 		return `(${input.toString()})()`;
 	},
-	'getItemFromCompendium': async function _getItemFromCompendium(key, name, ignoreNotFound) {
+	'getItemFromCompendium': async function _getItemFromCompendium(key, name, ignoreNotFound, packFolderId) {
 		let gamePack = game.packs.get(key);
 		if (!gamePack) {
 			ui.notifications.warn('Invalid compendium specified!');
 			return false;
 		}
 		let packItems = await gamePack.getDocuments();
-		let itemData = packItems.find(item => item.name === name);
+		let itemData;
+		if (!packFolderId) {
+			itemData = packItems.find(item => item.name === name);
+		} else {
+			itemData = packItems.find(item => item.name === name && item.flags.cf?.id === packFolderId)
+		}
 		if (!itemData) {
 			if (!ignoreNotFound) ui.notifications.warn('Item not found in specified compendium! Check spelling?');
 			return false;