+++ scripts/macros/monsterFeatures/generic/grapple.js
@@ -1,1 +1,5 @@
 import {chris} from '../../../helperFunctions.js';
+export async function autoGrapple(workflow) {
+    if (workflow.hitTargets.size != 1) return;
+    await chris.addCondition(workflow.targets.first().actor, 'Grappled', false, workflow.item.uuid);
+}