+++ scripts/macros/monsterFeatures/homunculus/homunculus.js
@@ -1,0 +1,4 @@
+import {bite} from './bite.js';
+export let homunculus = {
+    'bite': bite
+}