+++ scripts/macros/monsterFeatures/shadow/shadow.js
@@ -1,0 +1,4 @@
+import {strengthDrain} from './strengthDrain.js';
+export let shadow = {
+    'strengthDrain': strengthDrain
+}