+++ scripts/macros/monsterFeatures/shadowDemon/shadowDemon.js
@@ -1,0 +1,4 @@
+import {claws} from './claws.js';
+export let shadowDemon = {
+    'claws': claws
+}