+++ scripts/macros/monsterFeatures/succubus/succubus.js
@@ -1,0 +1,4 @@
+import {drainingKiss} from './drainingKiss.js';
+export let succubus = {
+    'drainingKiss': drainingKiss
+}